requirements.txt
Django
django-polymorphic